requirements.txt
amqp==1.0.11
anyjson==0.3.3
APScheduler==3.1.0
beautifulsoup4==4.3.2
billiard==2.7.3.28
blinker==1.3
boto==2.38.0
botocore==1.3.11
braintree==3.24.0
celery==3.0.19
click==5.1
dj-database-url==0.2.1
dj-static==0.0.5
Django==1.5.4
django-appconf==0.6
django-bitfield==1.6.4
django-celery==3.0.17
django-cors-headers==0.12
django-extensions==1.2.5
django-fields==0.3.0
django-filter==0.7
django-gcm==1.0.0
django-grappelli==2.6.5
-e git+https://github.com/stephenmuss/django-ios-notifications.git@9b3832d3c740798d0752fbc54ecea878069d4b05#egg=django_ios_notifications
django-redis-cache==0.10.0
django-rest-swagger==0.1.14
django-rq==0.5.1
django-storages==1.1.8
django-suit==0.2.7
django-tastypie==0.11.0
django-websocket-redis==0.4.1
djangorestframework==2.3.4
docutils==0.12
-e git+https://github.com/mobolic/facebook-sdk.git@c01be26406f8acd146363be168e4469ba6cf0a8a#egg=facebook_sdk
facebookads==2.8.1
foursquare==2015.2.2
funcsigs==1.0.0
futures==3.0.5
fuzzywuzzy==0.6.0
gevent==1.0.1
googlemaps==1.0.2
greenlet==0.4.2
gunicorn==0.17.4
httplib2==0.9
jmespath==0.9.0
jsonpickle==0.9.2
kombu==2.5.10
Markdown==2.4
mechanize==0.2.5
mixpanel-py==3.0.0
newrelic==2.4.0.4
numpy==1.8.0
oauth2==1.5.211
oauthlib==1.0.3
ordereddict==1.1
psycopg2==2.5
pycrypto==2.6
pygeocoder==1.2.2
pygeoip==0.3.1
pyOpenSSL==0.13
pystache==0.5.4
python-dateutil==2.1
python-instagram==1.3.2
python-mimeparse==0.1.4
python-twitter==2.2
pytz==2016.3
PyYAML==3.11
raven==3.6.1
raygun4py==3.1.1
readline==6.2.4.1
redis==2.10.3
requests==2.11.1
requests-oauthlib==0.5.0
rq==0.5.6
rq-scheduler==0.5.1
selenium==2.53.6
sendgrid==0.1.4
sendgrid-python==0.1.1
simplejson==3.8.0
six==1.10.0
sorl-thumbnail==11.12
South==0.8.1
static==1.0.2
times==0.6.2
twilio==3.7.3
tzlocal==1.2.2
unittest2==0.5.1
uWSGI==2.0.4
virtualenv==1.11.4
waitress==0.8.9
xmltodict==0.8.7
yelpapi==1.3
